devops-agent/agent/pyproject.toml
[project]
name = "terraform-agentic-orchestrator"
version = "0.1.0"
description = "Agent Framework powered Terraform automation orchestrator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-core>=0.1.0",
    "agent-framework-devui>=0.1.0",
    "agent-framework-ag-ui>=0.1.0",
    "azure-identity>=1.16.0",
    "fastapi>=0.111.0",
    "gitpython>=3.1.40",
    "httpx>=0.27.0",
    "langfuse>=2.0.0",
    "opentelemetry-exporter-otlp>=1.23.0",
    "opentelemetry-sdk>=1.23.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0.25",
    "tenacity>=8.2.3",
    "typing-extensions>=4.10.0",
    "uvicorn[standard]>=0.27.0",
    "databases[sqlite]>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx[http2]>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx[http2]>=0.27.0",
]

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"
